requirements.txt
awscli
boto3==1.21.14
click==8.0.4
dateutils==0.6.12
pytest==5.0.1
pytest-cov==2.7.1
s3fs==0.4.2
botocore==1.24.14
numpy==1.21.5
ordered-set==4.0.2
pandas==1.3.5
urllib3==1.26.8
pyarrow==6.0.1
awswrangler==2.14.0
pandasql==0.7.3
SQLAlchemy==1.4.32